+++ database/migrations/2022_09_30_000325_create_bl_bitacoras_libras_table.php
@@ -14,9 +14,8 @@ class CreateBlBitacorasLibrasTable extends Migration
     public function up()
     {
         Schema::create('bl_bitacoras_libras', function (Blueprint $table) {
-            $table->id();
-            $table->integer('bitacora_id');
-            $table->integer('libra_id');
+            $table->id();$table->unsignedBigInteger('bitacora_id');
+            $table->unsignedBigInteger('libra_id');
             $table->timestamps();
 
             $table->foreign('bitacora_id')->references('id')->on('bl_bitacoras')
